requirements.txt
numpy==2.4.0
pillow==12.0.0
